refactor: organize packages
