+++ app/src/main/java/com/voyager/data/mapper/RideConfirmMapper.kt
@@ -1,0 +1,10 @@
+package com.voyager.data.mapper
+
+import com.voyager.data.remote.model.ride.response.confirm.RideConfirmResponse
+import com.voyager.domain.model.ride.confirm.RideConfirm
+
+fun RideConfirmResponse.toDomain(): RideConfirm {
+    return RideConfirm(
+        success = success
+    )
+}
