--- app/src/main/java/com/voyager/data/model/mapper/RideConfirmMapper.kt
@@ -1,10 +1,0 @@
-package com.voyager.data.model.mapper
-
-import com.voyager.data.remote.model.ride_confirm_response.RideConfirmResponse
-import com.voyager.domain.model.ride.confirm.RideConfirm
-
-fun RideConfirmResponse.toDomain(): RideConfirm {
-    return RideConfirm(
-        success = success
-    )
-}
