--- app/src/main/java/com/voyager/data/remote/model/request_body/RideEstimateRequestBody.kt
@@ -1,15 +1,0 @@
-package com.voyager.data.remote.model.request_body
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RideEstimateRequestBody(
-
-    @SerialName("customer_id")
-    val customerId: String? = "",
-    @SerialName("origin")
-    val origin: String? = "",
-    @SerialName("destination")
-    val destination: String? = ""
-)
