+++ app/src/main/java/com/voyager/data/remote/model/ride/response/estimate/AddressResponse.kt
@@ -1,0 +1,12 @@
+package com.voyager.data.remote.model.ride.response.estimate
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AddressResponse (
+    @SerialName("latitude")
+    val latitude: Double? = null,
+    @SerialName("longitude")
+    val longitude: Double? = null
+)
