+++ app/src/main/java/com/voyager/data/remote/model/ride/response/estimate/RideEstimateErrorResponse.kt
@@ -1,0 +1,12 @@
+package com.voyager.data.remote.model.ride.response.estimate
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RideEstimateErrorResponse (
+    @SerialName("error_code")
+    val errorCode: String? = null,
+    @SerialName("error_description")
+    val errorDescription: String? = null
+)
