+++ app/src/main/java/com/voyager/data/remote/repository/RideRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package com.voyager.data.remote.repository
 
 import com.voyager.core.network.api.ApiService
-import com.voyager.data.remote.model.request_body.RideConfirmRequestBody
-import com.voyager.data.remote.model.request_body.RideEstimateRequestBody
-import com.voyager.data.remote.model.ride_confirm_response.RideConfirmResponse
-import com.voyager.data.remote.model.ride_estimate_response.RideEstimateResponse
+import com.voyager.data.remote.model.ride.request.confirm.RideConfirmRequestBody
+import com.voyager.data.remote.model.ride.request.estimate.RideEstimateRequestBody
+import com.voyager.data.remote.model.ride.response.confirm.RideConfirmResponse
+import com.voyager.data.remote.model.ride.response.estimate.RideEstimateResponse
 import com.voyager.domain.repository.RideRepository
 
 class RideRepositoryImpl(
