+++ app/src/main/java/com/voyager/dependency_injection/PresenterModule.kt
@@ -1,8 +1,8 @@
 package com.voyager.dependency_injection
 
-import com.voyager.presenter.screen.ride_history.viewModel.RideHistoryViewModel
-import com.voyager.presenter.screen.ride_confirm.view_model.RideConfirmViewModel
-import com.voyager.presenter.screen.ride_request.viewModel.RideRequestViewModel
+import com.voyager.presenter.screen.ride.history.viewModel.RideHistoryViewModel
+import com.voyager.presenter.screen.ride.confirm.view_model.RideConfirmViewModel
+import com.voyager.presenter.screen.ride.request.viewModel.RideRequestViewModel
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
