+++ app/src/main/java/com/voyager/domain/model/ride/estimate/Address.kt
@@ -1,0 +1,9 @@
+package com.voyager.domain.model.ride.estimate
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Address (
+    val latitude: Double? = null,
+    val longitude: Double? = null
+)
