+++ app/src/main/java/com/voyager/domain/model/ride/estimate/Review.kt
@@ -1,0 +1,9 @@
+package com.voyager.domain.model.ride.estimate
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Review (
+    val rating: Double? = null,
+    val comment: String? = null
+)
