+++ app/src/main/java/com/voyager/domain/repository/RideRepository.kt
@@ -1,9 +1,9 @@
 package com.voyager.domain.repository
 
-import com.voyager.data.remote.model.request_body.RideConfirmRequestBody
-import com.voyager.data.remote.model.request_body.RideEstimateRequestBody
-import com.voyager.data.remote.model.ride_confirm_response.RideConfirmResponse
-import com.voyager.data.remote.model.ride_estimate_response.RideEstimateResponse
+import com.voyager.data.remote.model.ride.request.confirm.RideConfirmRequestBody
+import com.voyager.data.remote.model.ride.request.estimate.RideEstimateRequestBody
+import com.voyager.data.remote.model.ride.response.confirm.RideConfirmResponse
+import com.voyager.data.remote.model.ride.response.estimate.RideEstimateResponse
 
 interface RideRepository {
