--- app/src/main/java/com/voyager/presenter/components/DefaultButton.kt
@@ -1,87 +1,0 @@
-package com.voyager.presenter.components
-
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.airbnb.lottie.compose.LottieAnimation
-import com.airbnb.lottie.compose.LottieCompositionSpec
-import com.airbnb.lottie.compose.LottieConstants
-import com.airbnb.lottie.compose.rememberLottieComposition
-import com.voyager.ui.theme.VoyagerTheme
-
-@Composable
-fun DefaultButton(
-    modifier: Modifier = Modifier,
-    text: String,
-    isLoading: Boolean = false,
-    enabled: Boolean = true,
-    onClick: () -> Unit = {}
-) {
-
-    val composition by rememberLottieComposition(LottieCompositionSpec.Asset("loading.json"))
-
-    Button(
-        onClick = onClick,
-        modifier = modifier
-            .fillMaxWidth()
-            .height(58.dp)
-            .shadow(
-                shape = RoundedCornerShape(100.dp),
-                elevation = 4.dp,
-                spotColor = MaterialTheme.colorScheme.primary,
-                ambientColor = MaterialTheme.colorScheme.primary
-            ),
-        enabled = enabled && !isLoading,
-        colors = ButtonDefaults.buttonColors(
-            containerColor = MaterialTheme.colorScheme.primary,
-            disabledContainerColor = MaterialTheme.colorScheme.secondary
-        ),
-        content = {
-            if (isLoading) {
-                LottieAnimation(
-                    composition = composition,
-                    modifier = Modifier
-                        .size(70.dp),
-                    iterations = LottieConstants.IterateForever,
-                    maintainOriginalImageBounds = true,
-                    contentScale = ContentScale.Crop
-                )
-            } else {
-                Text(
-                    text = text,
-                    style = TextStyle(
-                        fontSize = 18.sp,
-                        lineHeight = 22.4.sp,
-                        fontWeight = FontWeight.Bold,
-                        textAlign = TextAlign.Center,
-                        letterSpacing = 0.2.sp
-                    )
-                )
-            }
-        }
-    )
-}
-
-@PreviewLightDark
-@Composable
-fun DefaultButtonPreview() {
-    VoyagerTheme {
-        DefaultButton(text = "Button")
-    }
-}
