+++ app/src/main/java/com/voyager/presenter/screen/ride/confirm/action/RideConfirmAction.kt
@@ -1,0 +1,6 @@
+package com.voyager.presenter.screen.ride.confirm.action
+
+
+sealed class RideConfirmAction(
+
+)
