+++ app/src/main/java/com/voyager/presenter/screen/ride/confirm/state/RideConfirmState.kt
@@ -1,0 +1,9 @@
+package com.voyager.presenter.screen.ride.confirm.state
+
+import com.voyager.domain.model.ride.estimate.Option
+
+
+data class RideConfirmState(
+    val isLoading: Boolean = false,
+    val options: List<Option> = emptyList()
+)
