+++ app/src/main/java/com/voyager/presenter/screen/ride/history/action/RideHistoryAction.kt
@@ -1,0 +1,6 @@
+package com.voyager.presenter.screen.ride.history.action
+
+
+sealed class RideHistoryAction(
+
+)
