+++ app/src/main/java/com/voyager/presenter/screen/ride/history/state/RideHistoryState.kt
@@ -1,0 +1,6 @@
+package com.voyager.presenter.screen.ride.history.state
+
+
+data class RideHistoryState(
+    val isLoading: Boolean = true
+)
