--- app/src/main/java/com/voyager/presenter/screen/ride_confirm/action/RideConfirmAction.kt
@@ -1,6 +1,0 @@
-package com.voyager.presenter.screen.ride_confirm.action
-
-
-sealed class RideConfirmAction(
-
-)
