--- app/src/main/java/com/voyager/presenter/screen/ride_confirm/state/RideConfirmState.kt
@@ -1,6 +1,0 @@
-package com.voyager.presenter.screen.ride_confirm.state
-
-
-data class RideConfirmState(
-    val isLoading: Boolean = true
-)
