--- app/src/main/java/com/voyager/presenter/screen/ride_history/action/RideHistoryAction.kt
@@ -1,6 +1,0 @@
-package com.voyager.presenter.screen.ride_history.action
-
-
-sealed class RideHistoryAction(
-
-)
