--- app/src/main/java/com/voyager/presenter/screen/ride_history/state/RideHistoryState.kt
@@ -1,6 +1,0 @@
-package com.voyager.presenter.screen.ride_history.state
-
-
-data class RideHistoryState(
-    val isLoading: Boolean = true
-)
